Process a group of fields

This function calls other functions of the library to perform interpolation over different fields.

guinterp_process(
  inputpts,
  inlayer,
  filtered = TRUE,
  id_fieldname = "idfield",
  interp_dir = tempdir(),
  out_path = tempfile(),
  interp_method = "krige",
  smooth = TRUE,
  interp_res = 5,
  out_crs = NA,
  grid_offset = c(0, 0),
  samplesize = 10000,
  samplescheme = "random",
  buffer_radius = 15,
  parallel = TRUE,
  vgm = NA,
  v_nmax = Inf,
  v_maxdist = Inf,
  merge = TRUE,
  overwrite = TRUE,
  .shiny_session = NULL,
  .shiny_pbar_id = NULL
)

Arguments

inputpts
TODO

inlayer
SpatialPolygonsDataFrame of fields (see also id_fieldname)

filtered
logical: is inputpts already filtered? If TRUE (default), no additional filter is applied; if FALSE, a standard automatic filter is applied.

id_fieldname
optional: name of the inlayer field containing unique ID of fields (default: "idfield")

interp_dir
directory where rasters of single polygons are stores (default: temporary directory)

out_path
path of the output merged raster (default: temporary directory)

interp_method
interpolation method ("krige" or "idw")

smooth
logical: should rasters be smoothed after interpolation?

interp_res
output raster resolution (numeric, in metres)

out_crs
optional: CRS of output raster (default: CRS of inputpts)

grid_offset
optional: grid offset from point (0,0) in out_crs (2-length numeric)

samplesize
maximum size of the sample of the original data to work with (default: 10000; if NA: all the points)

samplescheme
sampling scheme within geometries ("random", "strat_npts" or "strat_area")

buffer_radius
numeric: buffer (default: 15) to be applied internally to field borders.

parallel
(optional) Logical or integer: if TRUE (default), interpolation is executed using multiple cores in order to speed up the execution. The number of cores is automatically determined; specifying it is also possible (e.g. parallel = 4). If FALSE, the interpolation is forced to run with a single core.

vgm
named list of variograms (names must correspond to values of field id_fieldname). If NA (default), they are automatically computed.

v_nmax
TODO.

v_maxdist
TODO.

merge
logical: if TRUE (default), merge raster of single fields and return the path of the merged raster; if FALSE, return the paths of the single field rasters.

overwrite
Logical: should existing raster being reprocessed? (default: FALSE)

.shiny_session
TODO

.shiny_pbar_id
TODO

Note

License: GPL 3.0