Convert to HTML entities

Internal functions used to convert elements to HTML.

t() is converts a list of character vectors to HTML.

ht() converts a vector of translation labels to a list of HTML translated entities.

ph() pastes HTML entities.

h(t)

ht(x, i18n)

ph(...)

Arguments

t
Character vector or list

x
Character vector or list

i18n
Object containing translations

...
Elements to be pasted.

Value

t() and ht(): if x is a 1-length character, an HTML object; if x is a vector with length > 1, a list of HTML objects; if x is a list of vectors, a list of objects.

ph(): an HTML element.

Note

License: GPL 3.0